Implement working chunkbuffer reading and writing
Yay! The chunkbuffer now supports writing and reading from it. It does
so in a simple, but efficient way. This is possible by use of existing
data structures (sharedbuffer, rollingreader) and by supporting parallel
reading and writing.

There are a few kinks to be worked out, but this is a good start. Right
now there is an end chunk for every worker. This may be remedied by
separating the Chunk method of Pile into ReaderChunk and WriterChunk.